Rename distance.py to similarity.py
